extracting the example
