+++ src/main/java/Adapter/AdvancedMediaPlayer/AdvancedMusicPlayer.java
@@ -1,4 +1,4 @@
-package Adapter.Ejemplo1.AdvancedMediaPlayer;
+package Adapter.AdvancedMediaPlayer;
 
 public interface AdvancedMusicPlayer {
     void playFullScreen(String fileName);
