+++ src/main/java/Adapter/MediaPlayer/MusicFormats.java
@@ -1,4 +1,4 @@
-package Adapter.Ejemplo1.MediaPlayer;
+package Adapter.MediaPlayer;
 
 public enum MusicFormats {
     MP3,
