--- src/main/java/Builder/cars/CarType.java
@@ -1,5 +1,0 @@
-package Builder.cars;
-
-public enum CarType {
-	CITY_CAR, SPORTS_CAR, SUV
-}
