--- src/main/java/Builder/components/Engine.java
@@ -1,43 +1,0 @@
-package Builder.components;
-
-/**
- * Just another feature of a car.
- */
-public class Engine {
-	private final double volume;
-	private double mileage;
-	private boolean started;
-
-	public Engine(double volume, double mileage) {
-		this.volume = volume;
-		this.mileage = mileage;
-	}
-
-	public void on() {
-		started = true;
-	}
-
-	public void off() {
-		started = false;
-	}
-
-	public boolean isStarted() {
-		return started;
-	}
-
-	public void go(double mileage) {
-		if (started) {
-			this.mileage += mileage;
-		} else {
-			System.err.println("Cannot go(), you must start engine first!");
-		}
-	}
-
-	public double getVolume() {
-		return volume;
-	}
-
-	public double getMileage() {
-		return mileage;
-	}
-}
