--- src/main/java/Builder/components/Transmission.java
@@ -1,8 +1,0 @@
-package Builder.components;
-
-/**
- * Just another feature of a car.
- */
-public enum Transmission {
-    SINGLE_SPEED, MANUAL, AUTOMATIC, SEMI_AUTOMATIC
-}
